fix: code refactoring into different services and utils
