+++ backend/src/main/java/com/example/webppt/model/ElementType.java
@@ -1,5 +1,5 @@
 package com.example.webppt.model;
 
 public enum ElementType {
-    TEXT, IMAGE, SHAPE, TABLE, CONNECTOR
+    TEXT, IMAGE, SHAPE, TABLE
 }
